// web/src/generated/backend/models/V1Backend.ts
/* tslint:disable */
/* eslint-disable */
/**
 * backend/v1/backend.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { V1BackendConfig } from './V1BackendConfig';
import {
    V1BackendConfigFromJSON,
    V1BackendConfigFromJSONTyped,
    V1BackendConfigToJSON,
    V1BackendConfigToJSONTyped,
} from './V1BackendConfig';
import type { V1Meta } from './V1Meta';
import {
    V1MetaFromJSON,
    V1MetaFromJSONTyped,
    V1MetaToJSON,
    V1MetaToJSONTyped,
} from './V1Meta';
import type { V1BackendStatus } from './V1BackendStatus';
import {
    V1BackendStatusFromJSON,
    V1BackendStatusFromJSONTyped,
    V1BackendStatusToJSON,
    V1BackendStatusToJSONTyped,
} from './V1BackendStatus';

/**
 * 
 * @export
 * @interface V1Backend
 */
export interface V1Backend {
    /**
     * 
     * @type {string}
     * @memberof V1Backend
     */
    version?: string;
    /**
     * 
     * @type {V1Meta}
     * @memberof V1Backend
     */
    meta?: V1Meta;
    /**
     * 
     * @type {V1BackendConfig}
     * @memberof V1Backend
     */
    config?: V1BackendConfig;
    /**
     * 
     * @type {V1BackendStatus}
     * @memberof V1Backend
     */
    status?: V1BackendStatus;
}

/**
 * Check if a given object implements the V1Backend interface.
 */
export function instanceOfV1Backend(value: object): value is V1Backend {
    return true;
}

export function V1BackendFromJSON(json: any): V1Backend {
    return V1BackendFromJSONTyped(json, false);
}

export function V1BackendFromJSONTyped(json: any, ignoreDiscriminator: boolean): V1Backend {
    if (json == null) {
        return json;
    }
    return {
        
        'version': json['version'] == null ? undefined : json['version'],
        'meta': json['meta'] == null ? undefined : V1MetaFromJSON(json['meta']),
        'config': json['config'] == null ? undefined : V1BackendConfigFromJSON(json['config']),
        'status': json['status'] == null ? undefined : V1BackendStatusFromJSON(json['status']),
    };
}

export function V1BackendToJSON(json: any): V1Backend {
    return V1BackendToJSONTyped(json, false);
}

export function V1BackendToJSONTyped(value?: V1Backend | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'version': value['version'],
        'meta': V1MetaToJSON(value['meta']),
        'config': V1BackendConfigToJSON(value['config']),
        'status': V1BackendStatusToJSON(value['status']),
    };
}